requirements.txt
asgiref==3.5.2
Django==4.0.5
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
gunicorn
mysqlclient
pytz==2022.1
sqlparse==0.4.2
